[OPIK-813] move AddToDatasetDialog to pages-shared folder
